Разбил проекты по отдельным папкам
