--- BousSOle.csproj
@@ -1,37 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-    <PropertyGroup>
-        <TargetFramework>net7.0</TargetFramework>
-        <Nullable>enable</Nullable>
-        <ImplicitUsings>enable</ImplicitUsings>
-        <DockerDefaultTargetOS>Linux</DockerDefaultTargetOS>
-    </PropertyGroup>
-
-    <ItemGroup>
-        <PackageReference Include="Microsoft.AspNetCore.OpenApi" Version="7.0.5" />
-        <PackageReference Include="Swashbuckle.AspNetCore" Version="6.4.0" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Compile Remove="LSO\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <EmbeddedResource Remove="LSO\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <Content Remove="LSO\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <None Remove="LSO\**" />
-    </ItemGroup>
-
-    <ItemGroup>
-      <_ContentIncludedByDefault Remove="LSO\obj\LSO.csproj.nuget.dgspec.json" />
-      <_ContentIncludedByDefault Remove="LSO\obj\project.assets.json" />
-      <_ContentIncludedByDefault Remove="LSO\obj\project.packagespec.json" />
-    </ItemGroup>
-
-</Project>
