+++ BousSOle.sln
@@ -1,8 +1,8 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BousSOle", "BousSOle.csproj", "{4AABBB02-E9A2-43AE-A259-C61CF5AE639D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BousSOle", "BousSOle\BousSOle.csproj", "{ABFAC716-90EE-4EAA-9868-B4C439E367FF}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LSO", "LSO\LSO.csproj", "{086BB519-FB40-4E7B-849D-3EA8CAE1E220}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LSO", "LSO\LSO.csproj", "{02273C07-2DF8-492E-BD73-67FD3C4ED9FE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -10,13 +10,13 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4AABBB02-E9A2-43AE-A259-C61CF5AE639D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4AABBB02-E9A2-43AE-A259-C61CF5AE639D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4AABBB02-E9A2-43AE-A259-C61CF5AE639D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4AABBB02-E9A2-43AE-A259-C61CF5AE639D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{086BB519-FB40-4E7B-849D-3EA8CAE1E220}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{086BB519-FB40-4E7B-849D-3EA8CAE1E220}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{086BB519-FB40-4E7B-849D-3EA8CAE1E220}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{086BB519-FB40-4E7B-849D-3EA8CAE1E220}.Release|Any CPU.Build.0 = Release|Any CPU
+		{ABFAC716-90EE-4EAA-9868-B4C439E367FF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{ABFAC716-90EE-4EAA-9868-B4C439E367FF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{ABFAC716-90EE-4EAA-9868-B4C439E367FF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{ABFAC716-90EE-4EAA-9868-B4C439E367FF}.Release|Any CPU.Build.0 = Release|Any CPU
+		{02273C07-2DF8-492E-BD73-67FD3C4ED9FE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02273C07-2DF8-492E-BD73-67FD3C4ED9FE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02273C07-2DF8-492E-BD73-67FD3C4ED9FE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02273C07-2DF8-492E-BD73-67FD3C4ED9FE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal
